Create AbstractClassPropertyEvent add class to properties when property changes
