+++ actions/class.PropertiesAuthoring.php
@@ -551,7 +551,13 @@ class tao_actions_PropertiesAuthoring extends tao_actions_CommonModule
         }
 
         if (count($changedProperties) > 0) {
-            $this->getEventManager()->trigger(new ClassPropertiesChangedEvent($changedProperties));
+            $this->getEventManager()->trigger(
+                new ClassPropertiesChangedEvent(
+                    array_merge($changedProperties, [
+                    'class' => $this->getCurrentClass()
+                    ])
+                )
+            );
         }
     }
